[AIG] Fix int64_t, size_t warning
Fix a compiler warning stemming from comparison of `size_t` to `int64_t`.
Opt for `size_t` here as this seems to be what is used everywhere else.

+++ lib/Dialect/AIG/Analysis/LongestPathAnalysis.cpp
@@ -77,7 +77,7 @@
 using namespace circt;
 using namespace aig;
 
-static int64_t getBitWidth(Value value) {
+static size_t getBitWidth(Value value) {
   if (auto vecType = dyn_cast<seq::ClockType>(value.getType()))
     return 1;
   if (auto memory = dyn_cast<seq::FirMemType>(value.getType()))
